feat: #27 SelectSpotNameFeature 이전
